requirements.txt
playwright>=1.56.0
python-dotenv>=1.2.0
aiohttp>=3.8.0
